+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `nixify` and `flakeify` from `nix-direnv`
 - Added inline colours to `VSCode`
 - Moved out of tree `vscode-neovim` extensions into their own overlay that magically merges them into `vscode-extensions`
 - Added editor plugins to visualize and trim trailing whitespace